Implementacion de Chain of Responsability en clase Bot
