+++ backend/src/asistente_virtual/IDecision.java
@@ -1,0 +1,8 @@
+package asistente_virtual;
+
+public interface IDecision {
+
+	public String leerMensaje(String mensaje, String usuario);
+	public IDecision getSiguienteDecision();
+	public void setSiguienteDecision(IDecision decision);
+}
