+++ backend/src/conversor_unidades/ConversorUnidad.java
@@ -1,8 +1,52 @@
 package conversor_unidades;
 
 import java.lang.reflect.Method;
+import java.text.DecimalFormat;
+import java.util.HashMap;
+import java.util.regex.Pattern;
 
-public class ConversorUnidad {
+import asistente_virtual.Bot;
+import asistente_virtual.IDecision;
+
+public class ConversorUnidad implements IDecision {
+	
+	private IDecision siguienteDecision;
+	
+	
+	@Override
+	public String leerMensaje(String mensaje, String usuario) {
+		if(mensaje.matches("@(\\w*) (?:cuantas|cuantos) (\\w*\\.*) (?:son|hay en) (\\d*\\.*\\d) (\\w+)(?:\\s* \\?|\\.*\\s*\\?)?")) {
+			mensaje = mensaje.replace("?", "");
+			String respuesta = "";
+			String[] palabras = mensaje.split(" ");
+			String hasta = palabras[2];
+			String desde = palabras[palabras.length - 1];
+	
+			Pattern formato_numero = Pattern.compile("(\\s+\\d*\\.*\\d)");
+			DecimalFormat df = new DecimalFormat("#0.00");
+			double numero = Bot.obtenerNumero(mensaje, formato_numero);
+			
+			ConversorUnidad cu = new ConversorUnidad();
+
+			double resultado = cu.convertirUnidad(numero, diccionario(desde), diccionario(hasta));
+			
+			if(resultado == -1) return Bot.MSG_NO_ENTIENDO;
+			if(resultado == -2) return "@" + usuario + " las magnitudes no pueden ser negativas.";
+			return respuesta = "@" + usuario + " " + df.format(numero) + " " + desde + 
+						" equivale a " + df.format(resultado) + " " + hasta;				
+		}
+		return  siguienteDecision.leerMensaje(mensaje, usuario);
+	}
+
+	@Override
+	public IDecision getSiguienteDecision() {
+		return siguienteDecision;
+	}
+
+	@Override
+	public void setSiguienteDecision(IDecision decision) {
+		siguienteDecision = decision;
+	}
 	
 	/**
 	 * @param numero: cantidad a convertir.
@@ -50,5 +94,101 @@ public class ConversorUnidad {
         return resultado;
     }
 
+	
+	/**
+	 * @param palara: palabra a evaluar.
+	 * @return string: retorna palabra asociada en el HashMap si existe.
+	 */
+	private String diccionario(String palabra) {
+	    String singular;
+	    HashMap<String,String> diccionario = new HashMap<>();
+	    diccionario.put("gr","Gramo");
+	    diccionario.put("gr.","Gramo");
+	    diccionario.put("grs","Gramo");
+	    diccionario.put("grs.","Gramo");
+	    diccionario.put("gramo","Gramo");
+	    diccionario.put("gramos","Gramo");
+	    diccionario.put("kilo","Kilo");
+	    diccionario.put("kilos","Kilo");
+	    diccionario.put("kg","Kilo");
+	    diccionario.put("kg.","Kilo");
+	    diccionario.put("kilogramo","Kilo");
+	    diccionario.put("kilogramos","Kilo");
+	    diccionario.put("onza","Onza");
+	    diccionario.put("onzas","Onza");
+	    diccionario.put("tonelada","Tonelada");
+	    diccionario.put("toneladas","Tonelada");
+	    diccionario.put("cc","Cm3");
+	    diccionario.put("cc.","Cm3");
+	    diccionario.put("cm3","Cm3");
+	    diccionario.put("cm3.","Cm3");
+	    diccionario.put("centimetro3","Cm3");
+	    diccionario.put("centimetros3","Cm3");
+	    diccionario.put("gal","Galon");
+	    diccionario.put("gal.","Galon");
+	    diccionario.put("galon","Galon");
+	    diccionario.put("galones","Galon");
+	    diccionario.put("lt","Litro");
+	    diccionario.put("lt.","Litro");
+	    diccionario.put("lts","Litro");
+	    diccionario.put("lts.","Litro");
+	    diccionario.put("litro","Litro");
+	    diccionario.put("litros","Litro");
+	    diccionario.put("mm","Milimetro");
+	    diccionario.put("mm.","Milimetro");
+	    diccionario.put("mms","Milimetro");
+	    diccionario.put("mms.","Milimetro");
+	    diccionario.put("milimetro","Milimetro");
+	    diccionario.put("milimetros","Milimetro");
+	    diccionario.put("cm","Centimetro");
+	    diccionario.put("cm.","Centimetro");
+	    diccionario.put("cms","Centimetro");
+	    diccionario.put("cms.","Centimetro");
+	    diccionario.put("centimetro","Centimetro");
+	    diccionario.put("centimetros","Centimetro");
+	    diccionario.put("mt","Metro");
+	    diccionario.put("mt.","Metro");
+	    diccionario.put("mts","Metro");
+	    diccionario.put("mts.","Metro");
+	    diccionario.put("metro","Metro");
+	    diccionario.put("metros","Metro");
+	    diccionario.put("km","Kilometro");
+	    diccionario.put("km.","Kilometro");
+	    diccionario.put("kms","Kilometro");
+	    diccionario.put("kms.","Kilometro");
+	    diccionario.put("kilometro","Kilometro");
+	    diccionario.put("kilometros","Kilometro");
+	    diccionario.put("pie","Pie");
+	    diccionario.put("pies","Pie");
+	    diccionario.put("plg","Pulgada");
+	    diccionario.put("plg.","Pulgada");
+	    diccionario.put("pulg","Pulgada");
+	    diccionario.put("pulg.","Pulgada");
+	    diccionario.put("pulgada","Pulgada");
+	    diccionario.put("pulgadas","Pulgada");
+	    diccionario.put("seg","Segundo");
+	    diccionario.put("seg.","Segundo");
+	    diccionario.put("segs","Segundo");
+	    diccionario.put("segs.","Segundo");
+	    diccionario.put("segundo","Segundo");
+	    diccionario.put("segundos","Segundo");
+	    diccionario.put("min","Minuto");
+	    diccionario.put("min.","Minuto");
+	    diccionario.put("mins","Minuto");
+	    diccionario.put("mins.","Minuto");
+	    diccionario.put("minuto","Minuto");
+	    diccionario.put("minutos","Minuto");
+	    diccionario.put("h","Hora");
+	    diccionario.put("h.","Hora");
+	    diccionario.put("hs","Hora");
+	    diccionario.put("hs.","Hora");
+	    diccionario.put("hora","Hora");
+	    diccionario.put("horas","Hora");
+	    diccionario.put("dia","Dia");
+	    diccionario.put("dias","Dia");
+	    singular = diccionario.get(palabra);
+	    return singular != null ? singular : palabra;
+	}
+	
 }
