+++ backend/tests/asistente_virtual/BotTests.java
@@ -2,6 +2,7 @@ package asistente_virtual;
 
 import static org.junit.Assert.assertEquals;
 
+import java.util.Calendar;
 import java.util.Date;
 import java.util.GregorianCalendar;
 
@@ -22,14 +23,77 @@ public class BotTests {
 	
 	@Test
 	public void mensajeInentendible() {
-		String rta = jenkins.leerMensaje("audwhkawud");
+		String rta = jenkins.leerMensaje("audwhkawud", USUARIO);
 		assertEquals(Bot.MSG_NO_ENTIENDO, rta);
 	}
 	
 	@Test
 	public void mensajeVacio() {
-		String rta = jenkins.leerMensaje("");
+		String rta = jenkins.leerMensaje("", USUARIO);
 		assertEquals(Bot.MSG_NO_ENTIENDO, rta);
+	}
+	
+	@Test
+	public void sinsentido() {
+		String[] mensajes = {
+				"Este mensaje no tiene sentido @jenkins"
+		};
+		for (String mensaje : mensajes) {
+			Assert.assertEquals(
+					"Disculpa... no entiendo el pedido, @delucas ¿podrás repetirlo?",
+					jenkins.leerMensaje(mensaje, USUARIO)
+			);
+		}
+	}
+	
+	@Test
+	public void saludo() {
+		Calendar calendario = Calendar.getInstance();
+		String[] mensajes = {
+				"¡Hola, @jenkins!",
+				"@jenkins hola!",
+				"buen día @jenkins",
+				"@jenkins, buenas tardes",
+				"hey @jenkins"
+		};
+		for (String mensaje : mensajes) {
+			if(calendario.get(Calendar.HOUR_OF_DAY) < 13)
+				Assert.assertEquals("Buenos días @" + USUARIO + "!", jenkins.leerMensaje(mensaje, USUARIO));
+			if(calendario.get(Calendar.HOUR_OF_DAY) > 19)
+				Assert.assertEquals("Buenas noches @" + USUARIO + "!", jenkins.leerMensaje(mensaje, USUARIO));
+			if(calendario.get(Calendar.HOUR_OF_DAY) >= 13 && calendario.get(Calendar.HOUR_OF_DAY) <= 19)
+				Assert.assertEquals("Buenas tardes @" + USUARIO + "!", jenkins.leerMensaje(mensaje, USUARIO));
+			}
+	}
+	
+	@Test
+	public void agradecimiento() {
+		String[] mensajes = {
+				"¡Muchas gracias, @jenkins!",
+				"@jenkins gracias",
+				"gracias @jenkins"
+		};
+		for (String mensaje : mensajes) {
+			Assert.assertEquals(
+					"No es nada, @" + Bot.USUARIO,
+					jenkins.leerMensaje(mensaje, USUARIO)
+			);
+		}
+	}
+	
+	@Test
+	public void despedida() {
+		String[] mensajes = {
+				"¡Chau @jenkins!",
+				"Adios @jenkins",
+				"Hasta luego @jenkins!"
+		};
+		for (String mensaje : mensajes) {
+			Assert.assertEquals(
+					"Hasta luego @" + Bot.USUARIO + "!",
+					jenkins.leerMensaje(mensaje, USUARIO)
+			);
+		}
 	}
 	
 	@Test
@@ -41,8 +105,8 @@ public class BotTests {
 		};
 		for (String mensaje : mensajes) {
 			Assert.assertEquals(
-					"@delucas son las 01:49 a. m.",
-					jenkins.leerMensaje(mensaje)
+					"@delucas son las 20:21 PM",
+					jenkins.leerMensaje(mensaje, USUARIO)
 			);
 		}
 	}
@@ -57,7 +121,7 @@ public class BotTests {
 		for (String mensaje : mensajes) {
 			Assert.assertEquals(
 					"@delucas hoy es 5 de mayo de 2018",
-					jenkins.leerMensaje(mensaje)
+					jenkins.leerMensaje(mensaje, USUARIO)
 			);
 		}
 	}
@@ -70,52 +134,7 @@ public class BotTests {
 		for (String mensaje : mensajes) {
 			Assert.assertEquals(
 					"@delucas hoy es sábado",
-					jenkins.leerMensaje(mensaje)
-			);
-		}
-	}
-	
-	@Test
-	public void sinsentido() {
-		String[] mensajes = {
-				"Este mensaje no tiene sentido @jenkins"
-		};
-		for (String mensaje : mensajes) {
-			Assert.assertEquals(
-					"Disculpa... no entiendo el pedido, @delucas ¿podrás repetirlo?",
-					jenkins.leerMensaje(mensaje)
-			);
-		}
-	}
-	
-	@Test
-	public void saludo() {
-		String[] mensajes = {
-				"¡Hola, @jenkins!",
-				"@jenkins hola!",
-				"buen día @jenkins",
-				"@jenkins, buenas tardes",
-				"hey @jenkins"
-		};
-		for (String mensaje : mensajes) {
-			Assert.assertEquals(
-					"¡Hola, @delucas!",
-					jenkins.leerMensaje(mensaje)
-			);
-		}
-	}
-	
-	@Test
-	public void agradecimiento() {
-		String[] mensajes = {
-				"¡Muchas gracias, @jenkins!",
-				"@jenkins gracias",
-				"gracias @jenkins"
-		};
-		for (String mensaje : mensajes) {
-			Assert.assertEquals(
-					"No es nada, @delucas",
-					jenkins.leerMensaje(mensaje)
+					jenkins.leerMensaje(mensaje, USUARIO)
 			);
 		}
 	}
@@ -124,17 +143,17 @@ public class BotTests {
 	public void diaDentroDe() {
 		Assert.assertEquals(
 				"@delucas será el domingo 20 de mayo de 2018",
-				jenkins.leerMensaje("@jenkins qué día será dentro de 2 días?")
+				jenkins.leerMensaje("@jenkins qué día será dentro de 2 días?", USUARIO)
 			);
 		
 		Assert.assertEquals(
 				"@delucas será el miércoles 18 de julio de 2018",
-				jenkins.leerMensaje("@jenkins qué día será dentro de 2 meses?")
+				jenkins.leerMensaje("@jenkins qué día será dentro de 2 meses?", USUARIO)
 			);
 		
 		Assert.assertEquals(
 				"@delucas será el lunes 18 de mayo de 2020",
-				jenkins.leerMensaje("@jenkins qué día será dentro de 2 años?")
+				jenkins.leerMensaje("@jenkins qué día será dentro de 2 años?", USUARIO)
 			);
 	}
 	
@@ -142,22 +161,22 @@ public class BotTests {
 	public void diaHace() {
 		Assert.assertEquals(
 				"@delucas fue viernes 4 de mayo de 2018",
-				jenkins.leerMensaje("@jenkins que dia fue ayer?")
+				jenkins.leerMensaje("@jenkins que dia fue ayer?", USUARIO)
 			);
 		
 		Assert.assertEquals(
 				"@delucas fue miércoles 2 de mayo de 2018",
-				jenkins.leerMensaje("@jenkins que dia fue hace 3 dias?")
+				jenkins.leerMensaje("@jenkins que dia fue hace 3 dias?", USUARIO)
 			);
 		
 		Assert.assertEquals(
 				"@delucas fue lunes 5 de marzo de 2018",
-				jenkins.leerMensaje("@jenkins que dia fue hace 2 meses?")
+				jenkins.leerMensaje("@jenkins que dia fue hace 2 meses?", USUARIO)
 			);
 		
 		Assert.assertEquals(
 				"@delucas fue jueves 5 de mayo de 2016",
-				jenkins.leerMensaje("@jenkins que dia fue hace 2 años?")
+				jenkins.leerMensaje("@jenkins que dia fue hace 2 años?", USUARIO)
 			);
 	}
 	
@@ -165,8 +184,8 @@ public class BotTests {
 	@Test
 	public void tiempoDesde() {
 		Assert.assertEquals(
-				"@delucas entre el 5 de mayo de 2017 y el 5 de mayo de 2018 pasaron 365 días",
-				jenkins.leerMensaje("@jenkins cuántos días pasaron desde el 5 de mayo de 2017?")
+				"@delucas entre el 5 de mayo de 2018 y el 24 de mayo de 2018 pasaron 19 días",
+				jenkins.leerMensaje("@jenkins cuántos días pasaron desde el 5 de mayo de 2018?", USUARIO)
 			);
 		
 		// agregar casos de prueba
@@ -176,7 +195,7 @@ public class BotTests {
 	public void tiempoHasta() {
 		Assert.assertEquals(
 				"@delucas faltan 1 días",
-				jenkins.leerMensaje("@jenkins cuántos días faltan para el 6 de mayo?")
+				jenkins.leerMensaje("@jenkins cuántos días faltan para el 6 de mayo?", USUARIO)
 			);
 		
 		// agregar casos de prueba
@@ -186,23 +205,23 @@ public class BotTests {
 	public void calculos() {
 		Assert.assertEquals(
 				"@delucas 3",
-				jenkins.leerMensaje("@jenkins cuánto es 1 + 2")
+				jenkins.leerMensaje("@jenkins cuánto es 1 + 2", USUARIO)
 			);
 		
 		
 		Assert.assertEquals(
 				"@delucas 1",
-				jenkins.leerMensaje("@jenkins cuánto es 5 - 2 * 2")
+				jenkins.leerMensaje("@jenkins cuánto es 5 - 2 * 2", USUARIO)
 			);
 		
 		Assert.assertEquals(
 				"@delucas 10",
-				jenkins.leerMensaje("@jenkins cuánto es el 10% de 100")
+				jenkins.leerMensaje("@jenkins cuánto es el 10% de 100", USUARIO)
 			);
 		
 		Assert.assertEquals(
 				"@delucas 42",
-				jenkins.leerMensaje("@jenkins cuánto es el 17 + 5 ^ 2")
+				jenkins.leerMensaje("@jenkins cuánto es el 17 + 5 ^ 2", USUARIO)
 			);
 		
 		// agregar otros casos
@@ -212,7 +231,7 @@ public class BotTests {
 	public void calculosCompuestos() {
 		Assert.assertEquals(
 				"@delucas -6",
-				jenkins.leerMensaje("@jenkins cuánto es (4-8)*2 + 4 / ( 1 + 1)")
+				jenkins.leerMensaje("@jenkins cuánto es (4-8)*2 + 4 / ( 1 + 1)", USUARIO)
 			);
 		
 		// agregar otros casos
@@ -225,7 +244,7 @@ public class BotTests {
 	public void magnitudInvalida() {
 		Assert.assertEquals(
 				Bot.MSG_NO_ENTIENDO,
-				jenkins.leerMensaje("@jenkins cuántos gramos son 1 rama")
+				jenkins.leerMensaje("@jenkins cuántos gramos son 1 rama", USUARIO)
 			);
 	}
 	
@@ -233,7 +252,7 @@ public class BotTests {
 	public void convertirCeroUnidad() {
 		Assert.assertEquals(
 				"@delucas 0,00 galones equivale a 0,00 litros",
-				jenkins.leerMensaje("@jenkins cuántos litros son 0 galones?")
+				jenkins.leerMensaje("@jenkins cuántos litros son 0 galones?", USUARIO)
 			);
 	}
 	
@@ -241,57 +260,57 @@ public class BotTests {
 	public void convertirUnidadRegex() {
 		Assert.assertEquals(
 				"@delucas 1,00 kilo equivale a 1000,00 gramos",
-				jenkins.leerMensaje("@jenkins cuántos gramos son 1 kilo?")
+				jenkins.leerMensaje("@jenkins cuántos gramos son 1 kilo?", USUARIO)
 			);
 		
 		Assert.assertEquals(
 				"@delucas 1,00 kilo equivale a 1000,00 gramos",
-				jenkins.leerMensaje("@jenkins cuántos gramos hay en 1 kilo?")
+				jenkins.leerMensaje("@jenkins cuántos gramos hay en 1 kilo?", USUARIO)
 			);
 
 		Assert.assertEquals(
 				"@delucas 1,00 kilo equivale a 1000,00 gramos",
-				jenkins.leerMensaje("@jenkins cuántas gramos son 1 kilo?")
+				jenkins.leerMensaje("@jenkins cuántas gramos son 1 kilo?", USUARIO)
 			);
 		
 		Assert.assertEquals(
 				"@delucas 1,00 kilo equivale a 1000,00 gramos",
-				jenkins.leerMensaje("@jenkins cuántas gramos hay en 1 kilo?")
+				jenkins.leerMensaje("@jenkins cuántas gramos hay en 1 kilo?", USUARIO)
 			);
 		
 		Assert.assertEquals(
 				"@delucas 1,00 kilo equivale a 1000,00 gramos",
-				jenkins.leerMensaje("@jenkins cuántos gramos son 1 kilo ?")
+				jenkins.leerMensaje("@jenkins cuántos gramos son 1 kilo ?", USUARIO)
 			);
 		
 		Assert.assertEquals(
 				"@delucas 1,00 kilo equivale a 1000,00 gramos",
-				jenkins.leerMensaje("@jenkins cuántos gramos son 1 kilo   ?")
+				jenkins.leerMensaje("@jenkins cuántos gramos son 1 kilo   ?", USUARIO)
 			);
 		
 		Assert.assertEquals(
 				"@delucas 100,00 mts. equivale a 10000,00 cms.",
-				jenkins.leerMensaje("@jenkins cuántos cms. son 100 mts.?")
+				jenkins.leerMensaje("@jenkins cuántos cms. son 100 mts.?", USUARIO)
 			);
 		
 		Assert.assertEquals(
 				"@delucas 100,00 mts equivale a 10000,00 cms.",
-				jenkins.leerMensaje("@jenkins cuántos cms. son 100 mts?")
+				jenkins.leerMensaje("@jenkins cuántos cms. son 100 mts?", USUARIO)
 			);
 		
 		Assert.assertEquals(
 				"@delucas 100,00 mts. equivale a 10000,00 cms",
-				jenkins.leerMensaje("@jenkins cuántos cms son 100 mts.?")
+				jenkins.leerMensaje("@jenkins cuántos cms son 100 mts.?", USUARIO)
 			);
 		
 		Assert.assertEquals(
 				"@delucas 100,00 mts. equivale a 10000,00 cms.",
-				jenkins.leerMensaje("@jenkins cuántos cms. son 100 mts. ?")
+				jenkins.leerMensaje("@jenkins cuántos cms. son 100 mts. ?", USUARIO)
 			);
 		
 		Assert.assertEquals(
 				"@delucas 100,00 mts. equivale a 10000,00 cms.",
-				jenkins.leerMensaje("@jenkins cuántos cms. son 100 mts.  ?")
+				jenkins.leerMensaje("@jenkins cuántos cms. son 100 mts.  ?", USUARIO)
 			);
 		
 	}
@@ -301,37 +320,37 @@ public class BotTests {
 		
 		Assert.assertEquals(
 				"@delucas 1,00 kilo equivale a 1000,00 gramos",
-				jenkins.leerMensaje("@jenkins cuántos gramos son 1 kilo")
+				jenkins.leerMensaje("@jenkins cuántos gramos son 1 kilo", USUARIO)
 			);
 		
 		Assert.assertEquals(
 				"@delucas 1,00 kilogramo equivale a 1000,00 gr",
-				jenkins.leerMensaje("@jenkins cuántos gr son 1 kilogramo")
+				jenkins.leerMensaje("@jenkins cuántos gr son 1 kilogramo", USUARIO)
 			);
 
 		Assert.assertEquals(
 				"@delucas 2,00 kilos equivale a 2000,00 gramos",
-				jenkins.leerMensaje("@jenkins cuántos gramos son 2 kilos")
+				jenkins.leerMensaje("@jenkins cuántos gramos son 2 kilos", USUARIO)
 			);
 		
 		Assert.assertEquals(
 				"@delucas 1000,00 gramos equivale a 1,00 kilos",
-				jenkins.leerMensaje("@jenkins cuántos kilos son 1000 gramos")
+				jenkins.leerMensaje("@jenkins cuántos kilos son 1000 gramos", USUARIO)
 			);
 		
 		Assert.assertEquals(
 				"@delucas 1000,00 gramos equivale a 35,27 onzas",
-				jenkins.leerMensaje("@jenkins cuántas onzas son 1000 gramos")
+				jenkins.leerMensaje("@jenkins cuántas onzas son 1000 gramos", USUARIO)
 			);
 		
 		Assert.assertEquals(
 				"@delucas 4000000,00 gramos equivale a 4,00 toneladas",
-				jenkins.leerMensaje("@jenkins cuántas toneladas son 4000000 gramos")
+				jenkins.leerMensaje("@jenkins cuántas toneladas son 4000000 gramos", USUARIO)
 			);
 		
 		Assert.assertEquals(
 				"@delucas 1,70 toneladas equivale a 1700,00 kilos",
-				jenkins.leerMensaje("@jenkins cuántos kilos son 1.7 toneladas")
+				jenkins.leerMensaje("@jenkins cuántos kilos son 1.7 toneladas", USUARIO)
 			);
 	}
 
@@ -339,22 +358,22 @@ public class BotTests {
 	public void unidadesDeTiempo() {
 		Assert.assertEquals(
 				"@delucas 60,00 segundos equivale a 1,00 minutos",
-				jenkins.leerMensaje("@jenkins cuántos minutos hay en 60 segundos")
+				jenkins.leerMensaje("@jenkins cuántos minutos hay en 60 segundos", USUARIO)
 			);
 
 		Assert.assertEquals(
 				"@delucas 3,50 horas equivale a 12600,00 segundos",
-				jenkins.leerMensaje("@jenkins cuántos segundos hay en 3.5 horas?")
+				jenkins.leerMensaje("@jenkins cuántos segundos hay en 3.5 horas?", USUARIO)
 			);
 		
 		Assert.assertEquals(
 				"@delucas 3,50 horas equivale a 12600,00 seg.",
-				jenkins.leerMensaje("@jenkins cuántas seg. hay en 3.5 horas?")
+				jenkins.leerMensaje("@jenkins cuántas seg. hay en 3.5 horas?", USUARIO)
 			);
 
 		Assert.assertEquals(
 				"@delucas 3,50 hs. equivale a 12600,00 segundo",
-				jenkins.leerMensaje("@jenkins cuántas segundo hay en 3.5 hs.?")
+				jenkins.leerMensaje("@jenkins cuántas segundo hay en 3.5 hs.?", USUARIO)
 			);
 	}
 	
@@ -362,17 +381,17 @@ public class BotTests {
 	public void unidadesDeCapacidad() {
 		Assert.assertEquals(
 				"@delucas 3500,00 cm3 equivale a 3,50 litros",
-				jenkins.leerMensaje("@jenkins cuántos litros hay en 3500 cm3")
+				jenkins.leerMensaje("@jenkins cuántos litros hay en 3500 cm3", USUARIO)
 			);
 
 		Assert.assertEquals(
 				"@delucas 10,00 galones equivale a 45,46 litros",
-				jenkins.leerMensaje("@jenkins cuántos litros son 10 galones?")
+				jenkins.leerMensaje("@jenkins cuántos litros son 10 galones?", USUARIO)
 			);
 		
 		Assert.assertEquals(
 				"@delucas 10,00 lts. equivale a 10000,00 cm3.",
-				jenkins.leerMensaje("@jenkins cuántos cm3. son 10 lts.?")
+				jenkins.leerMensaje("@jenkins cuántos cm3. son 10 lts.?", USUARIO)
 			);
 	}
 	
@@ -380,17 +399,17 @@ public class BotTests {
 	public void unidadesDeLongitud() {
 		Assert.assertEquals(
 				"@delucas 1000,00 metros equivale a 1,00 kilometros",
-				jenkins.leerMensaje("@jenkins cuántos kilometros hay en 1000 metros")
+				jenkins.leerMensaje("@jenkins cuántos kilometros hay en 1000 metros", USUARIO)
 			);
 
 		Assert.assertEquals(
 				"@delucas 1,00 pulgada equivale a 2,54 centimetros",
-				jenkins.leerMensaje("@jenkins cuántos centímetros son 1 pulgada?")
+				jenkins.leerMensaje("@jenkins cuántos centímetros son 1 pulgada?", USUARIO)
 			);
 		
 		Assert.assertEquals(
 				"@delucas 100,00 mts. equivale a 10000,00 cms",
-				jenkins.leerMensaje("@jenkins cuántos cms son 100 mts.?")
+				jenkins.leerMensaje("@jenkins cuántos cms son 100 mts.?", USUARIO)
 			);
 		
 	}
